Create Trading Algorithm - Regression Based
Trading algorithm, implemented with Alpaca Paper Trading API. Based on pairs trading, in this case uses MasterCard (MA) and Visa (V). The idea is that the prices for these two are correlated, and that the and that price changes in one are refelcted in the other but that on average they move together (if one shoots up, the other will dip). Based on this we can calculate a 'residual': the diference between actual current price and what is expected according to the regression. This residual changes with price and is mean reverting. We enter and exit trades based on how much the residual changes. If the current price is above the expected, ie current price > expected, the residual is positive and we can expect it to decrease, so we enter a short positon. If residual is under expected, actual price is under expected, and we enter a long position and wait for the residual to increase back to zero.